%do not edit, edit noweb/qmrparser.nw
\name{qmrparser-package}
\alias{qmrparser-package}
\alias{qmrparser}
\docType{package}
\title{
Parser Combinator in R
}
\description{
Basic functions for building parsers, with an application to PC-AXIS format files.
}
\details{
\tabular{ll}{
Package: \tab qmrparser\cr
Type:    \tab Package\cr
Version: \tab 0.1.6\cr
Date:    \tab 2022-04-10\cr
License: \tab GPL (>= 3)\cr
LazyLoad:\tab yes\cr
}
Collection of functions to build programs to read complex data files formats, with an application to the case of PC-AXIS format.
}
\author{
 Juan Gea Rosat, Ramon Martínez Coscollà 

Maintainer: Juan Gea Rosat <juangea@geax.net>
}
\references{

Parser combinator. 
\url{https://en.wikipedia.org/wiki/Parser_combinator}

Context-free grammar.
\url{https://en.wikipedia.org/wiki/Context-free_grammar}    

PC-Axis file format.
\url{https://www.scb.se/en/services/statistical-programs-for-px-files/px-file-format/}

Type \code{RShowDoc("index",package="qmrparser")} at the R command line to open the package vignette.

Type \code{ RShowDoc("qmrparser",package="qmrparser")} to open pdf developer guide.

Source code used in literate programming can be found in folder 'noweb'.
}
\keyword{ package }
\keyword{ parser combinator }
\keyword{ token   }
\keyword{ PC-AXIS }
